41: extract the symbols for API authentication into their own module
