+++ backend/src/__init__.py
@@ -1,4 +1,4 @@
-from src.vocab_treasury import app, db
+from src.vocab_treasury import app, db, flsk_bcrpt, token_serializer
 
 
-__all__ = [app, db]
+__all__ = [app, db, flsk_bcrpt, token_serializer]
