+++ backend/src/vocab_treasury.py
@@ -1,5 +1,4 @@
 from flask import Flask, request, jsonify, url_for
-from flask_httpauth import HTTPBasicAuth, HTTPTokenAuth
 from dotenv import find_dotenv, load_dotenv
 from flask_sqlalchemy import SQLAlchemy
 from flask_migrate import Migrate
@@ -85,87 +84,10 @@ mail = Mail(app)
 # As suggested on https://flask-bcrypt.readthedocs.io/en/latest/
 # the following variable is not called simply `bcrypt`
 # (because if it were, it would be effectively overriding the `bcrypt` module).
-flask_bcrypt = Bcrypt(app)
+flsk_bcrpt = Bcrypt(app)
 
 
 from src.models import User, Example
-
-
-basic_auth = HTTPBasicAuth()
-
-
-@basic_auth.verify_password
-def verify_password(email, password):
-    user = User.query.filter_by(email=email).first()
-    if user is None:
-        return None
-
-    if flask_bcrypt.check_password_hash(user.password_hash, password) is False:
-        return None
-
-    return user
-
-
-@basic_auth.error_handler
-def basic_auth_error():
-    """Return a 401 error to the client."""
-    r = jsonify(
-        {
-            "error": "Unauthorized",
-            "message": "Authentication in the Basic Auth format is required.",
-        }
-    )
-    r.status_code = 401
-    # fmt: off
-    '''
-    source: https://blog.miguelgrinberg.com/post/designing-a-restful-api-with-python-and-flask
-
-        Unfortunately web browsers have the nasty habit of showing an ugly login dialog
-        box when a request comes back with a 401 error code.
-
-        A simple trick to distract web browsers is to return an error code other than
-        401. An alternative error code favored by many is 403, which is the "Forbidden"
-        error. While this is a close enough error, it sort of violates the HTTP
-        standard, so it is not the proper thing to do if full compliance is necessary.
-        In particular this would be a bad idea if the client application is not a web
-        browser. But for cases where server and client are developed together it saves a
-        lot of trouble.
-    
-    source: https://flask-httpauth.readthedocs.io/en/latest/index.html
-
-        class flask_httpauth.HTTPBasicAuth
-
-            This class handles HTTP Basic authentication for Flask routes.
-
-            __init__(scheme=None, realm=None)
-
-                Create a basic authentication object.
-
-                If the optional `scheme` argument is provided, it will be used instead
-                of the standard “Basic” scheme in the `WWW-Authenticate` response. A
-                fairly common practice is to use a custom scheme to prevent browsers
-                from prompting the user to login.
-
-                The `realm` argument can be used to provide an application defined realm
-                with the `WWW-Authenticate` header.
-
-    source: https://github.com/kaloyan-marinov/goal-tracker-public/blob/13c76f8abd4182e8ad559492e796476eb9d38f37/goal_tracker/auth.py
-
-        This project also prevents web browsers from showing an ugly login dialog box
-        when a request comes back with a 401 error code.
-
-        It appears that:
-        (a) this project _effectively_ uses the approach recommended by the previous
-            source, but
-        (b) this project's implementation is not identical to the approach recommended
-            by the previous source.
-        
-        Recall that this project's implementation is based on the following project's
-        implementation:
-        https://github.com/miguelgrinberg/microblog/blob/02aae8d9816fb23b72cbeca2ac945ac2d48a6ed0/app/api/auth.py
-    '''
-    # fmt: on
-    return r
 
 
 token_serializer = TimedJSONWebSignatureSerializer(
@@ -181,36 +103,10 @@ token_serializer_for_password_resets = TimedJSONWebSignatureSerializer(
 )
 
 
-token_auth = HTTPTokenAuth()
-
-
-@token_auth.verify_token
-def verify_token(token):
-    try:
-        token_payload = token_serializer.loads(token)
-    except SignatureExpired:
-        return None  # valid token, but expired
-    except BadSignature:
-        return None  # invalid token
-
-    user = User.query.get(token_payload["user_id"])
-    if user is None:
-        return None
-
-    return user
-
-
-@token_auth.error_handler
-def token_auth_error():
-    """Return a 401 error to the client."""
-    r = jsonify(
-        {
-            "error": "Unauthorized",
-            "message": "Authentication in the Bearer-Token Auth format is required.",
-        }
-    )
-    r.status_code = 401
-    return r
+from src.auth import (
+    basic_auth,
+    token_auth,
+)
 
 
 @app.route("/api/users", methods=["POST"])
@@ -275,7 +171,7 @@ def create_user():
     user = User(
         username=username,
         email=email,
-        password_hash=flask_bcrypt.generate_password_hash(password).decode("utf-8"),
+        password_hash=flsk_bcrpt.generate_password_hash(password).decode("utf-8"),
     )
     db.session.add(user)
     db.session.commit()
@@ -405,7 +301,7 @@ def edit_user(user_id):
     if email is not None:
         u.email = email
     if password is not None:
-        u.password_hash = flask_bcrypt.generate_password_hash(password).decode("utf-8")
+        u.password_hash = flsk_bcrpt.generate_password_hash(password).decode("utf-8")
 
     db.session.add(u)
     db.session.commit()
@@ -591,7 +487,7 @@ def reset_password(token):
         r.status_code = 400
         return r
 
-    user.password_hash = flask_bcrypt.generate_password_hash(new_password)
+    user.password_hash = flsk_bcrpt.generate_password_hash(new_password)
     db.session.add(user)
     db.session.commit()
